// crowdfunding/src/types/FundProject.ts
export interface FundProjectInterface extends utils.Interface {
  functions: {
    "categories()": FunctionFragment;
    "description()": FunctionFragment;
    "founders()": FunctionFragment;
    "goalAmount()": FunctionFragment;
    "image()": FunctionFragment;
    "mail()": FunctionFragment;
    "makeDonation()": FunctionFragment;
    "raisedAmount()": FunctionFragment;
    "social()": FunctionFragment;
    "title()": FunctionFragment;
  };

  getFunction(
    nameOrSignatureOrTopic:
      | "categories"
      | "description"
      | "founders"
      | "goalAmount"
      | "image"
      | "mail"
      | "makeDonation"
      | "raisedAmount"
      | "social"
      | "title"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "categories",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "description",
    values?: undefined
  ): string;
  encodeFunctionData(functionFragment: "founders", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "goalAmount",
    values?: undefined
  ): string;
  encodeFunctionData(functionFragment: "image", values?: undefined): string;
  encodeFunctionData(functionFragment: "mail", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "makeDonation",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "raisedAmount",
    values?: undefined
  ): string;
  encodeFunctionData(functionFragment: "social", values?: undefined): string;
  encodeFunctionData(functionFragment: "title", values?: undefined): string;

  decodeFunctionResult(functionFragment: "categories", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "description",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "founders", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "goalAmount", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "image", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "mail", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "makeDonation",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "raisedAmount",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "social", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "title", data: BytesLike): Result;

  events: {
    "Funded(address,uint256,uint256)": EventFragment;
  };

  getEvent(nameOrSignatureOrTopic: "Funded"): EventFragment;
}

export interface FundedEventObject {
  donar: string;
  amount: BigNumber;
  timestamp: BigNumber;
}
export type FundedEvent = TypedEvent<
  [string, BigNumber, BigNumber],
  FundedEventObject
>;

export type FundedEventFilter = TypedEventFilter<FundedEvent>;

export interface FundProject extends BaseContract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: FundProjectInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    categories(overrides?: CallOverrides): Promise<[string]>;

    description(overrides?: CallOverrides): Promise<[string]>;

    founders(overrides?: CallOverrides): Promise<[string]>;

    goalAmount(overrides?: CallOverrides): Promise<[BigNumber]>;

    image(overrides?: CallOverrides): Promise<[string]>;

    mail(overrides?: CallOverrides): Promise<[string]>;

    makeDonation(
      overrides?: PayableOverrides & { from?: PromiseOrValue<string> }
    ): Promise<ContractTransaction>;

    raisedAmount(overrides?: CallOverrides): Promise<[BigNumber]>;

    social(overrides?: CallOverrides): Promise<[string]>;

    title(overrides?: CallOverrides): Promise<[string]>;
  };

  categories(overrides?: CallOverrides): Promise<string>;

  description(overrides?: CallOverrides): Promise<string>;

  founders(overrides?: CallOverrides): Promise<string>;

  goalAmount(overrides?: CallOverrides): Promise<BigNumber>;

  image(overrides?: CallOverrides): Promise<string>;

  mail(overrides?: CallOverrides): Promise<string>;

  makeDonation(
    overrides?: PayableOverrides & { from?: PromiseOrValue<string> }
  ): Promise<ContractTransaction>;

  raisedAmount(overrides?: CallOverrides): Promise<BigNumber>;

  social(overrides?: CallOverrides): Promise<string>;

  title(overrides?: CallOverrides): Promise<string>;

  callStatic: {
    categories(overrides?: CallOverrides): Promise<string>;

    description(overrides?: CallOverrides): Promise<string>;

    founders(overrides?: CallOverrides): Promise<string>;

    goalAmount(overrides?: CallOverrides): Promise<BigNumber>;

    image(overrides?: CallOverrides): Promise<string>;

    mail(overrides?: CallOverrides): Promise<string>;

    makeDonation(overrides?: CallOverrides): Promise<void>;

    raisedAmount(overrides?: CallOverrides): Promise<BigNumber>;

    social(overrides?: CallOverrides): Promise<string>;

    title(overrides?: CallOverrides): Promise<string>;
  };

  filters: {
    "Funded(address,uint256,uint256)"(
      donar?: PromiseOrValue<string> | null,
      amount?: PromiseOrValue<BigNumberish> | null,
      timestamp?: PromiseOrValue<BigNumberish> | null
    ): FundedEventFilter;
    Funded(
      donar?: PromiseOrValue<string> | null,
      amount?: PromiseOrValue<BigNumberish> | null,
      timestamp?: PromiseOrValue<BigNumberish> | null
    ): FundedEventFilter;
  };

  estimateGas: {
    categories(overrides?: CallOverrides): Promise<BigNumber>;

    description(overrides?: CallOverrides): Promise<BigNumber>;

    founders(overrides?: CallOverrides): Promise<BigNumber>;

    goalAmount(overrides?: CallOverrides): Promise<BigNumber>;

    image(overrides?: CallOverrides): Promise<BigNumber>;

    mail(overrides?: CallOverrides): Promise<BigNumber>;

    makeDonation(
      overrides?: PayableOverrides & { from?: PromiseOrValue<string> }
    ): Promise<BigNumber>;

    raisedAmount(overrides?: CallOverrides): Promise<BigNumber>;

    social(overrides?: CallOverrides): Promise<BigNumber>;

    title(overrides?: CallOverrides): Promise<BigNumber>;
  };

  populateTransaction: {
    categories(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    description(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    founders(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    goalAmount(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    image(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    mail(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    makeDonation(
      overrides?: PayableOverrides & { from?: PromiseOrValue<string> }
    ): Promise<PopulatedTransaction>;

    raisedAmount(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    social(overrides?: CallOverrides): Promise<PopulatedTransaction>;

    title(overrides?: CallOverrides): Promise<PopulatedTransaction>;
  };
}